+++ CHANGELOG.md
@@ -21,6 +21,10 @@ Types of changes:
 
 - `<ColorInput/>` - new component for HEX color input [#2898](https://github.com/wix/wix-style-react/pull/2898)
 
+### Docs
+
+- `<Page/>` (Docs): Make docs describe the New Page. Improve Component Doc Page. Add examples. [#3049](https://github.com/wix/wix-style-react/pull/3049)
+
 ### Deprecated
 
 - `<RadioGroup/>` - deprecated prop `type="button"` [#2874](https://github.com/wix/wix-style-react/pull/2874)